+++ test/hexapdf/content/test_canvas.rb
@@ -1080,6 +1080,21 @@ describe HexaPDF::Content::Canvas do
     end
   end
 
+  describe "show_glyphs_only" do
+    it "serializes correctly" do
+      @canvas.font("Times", size: 20)
+      font = @canvas.font
+      @canvas.show_glyphs_only(font.decode_utf8("Hal lo").insert(2, -35))
+      assert_equal(0, @canvas.text_cursor[0])
+      assert_equal(0, @canvas.text_cursor[1])
+      assert_operators(@canvas.contents, [[:set_font_and_size, [:F1, 20]],
+                                          [:set_leading, [24]],
+                                          [:begin_text],
+                                          [:show_text_with_positioning, [["Ha", -35, "l lo"]]],
+                                         ])
+    end
+  end
+
   describe "text" do
     it "sets the text cursor position if instructed" do
       @canvas.font("Times", size: 10)